@@ -1,5 +1,5 @@
 {
-    "name": "CommSy",
+    "name": "commsy/commsy",
     "license": "GPLv2",
     "type": "project",
     "autoload": {
@@ -19,47 +19,44 @@
     "require": {
         "php": "^7.1.3",
         "symfony/symfony": "3.4.*",
-        "doctrine/orm": "^2.5",
-        "doctrine/doctrine-bundle": "^1.6",
-        "doctrine/doctrine-cache-bundle": "^1.2",
-        "twig/extensions": "^1.3",
-        "symfony/swiftmailer-bundle": "^2.6.4",
-        "symfony/monolog-bundle": "^3.1.0",
-        "symfony/polyfill-apcu": "^1.0",
+        "doctrine/orm": "^2.6",
+        "doctrine/doctrine-bundle": "^1.10",
+        "twig/extensions": "^1.5",
+        "symfony/swiftmailer-bundle": "^3.2",
+        "symfony/monolog-bundle": "^3.3",
+        "symfony/polyfill-apcu": "^1.10",
         "sensio/distribution-bundle": "^5.0",
-        "sensio/framework-extra-bundle": "^5.0.0",
-        "incenteev/composer-parameter-handler": "^2.0",
-        "doctrine/doctrine-migrations-bundle": "^1.0",
-        "liip/theme-bundle": "^1.4",
-        "knplabs/knp-menu-bundle":"^2.1",
+        "sensio/framework-extra-bundle": "^5.2",
+        "incenteev/composer-parameter-handler": "^2.1",
+        "doctrine/doctrine-migrations-bundle": "^2.0",
+        "liip/theme-bundle": "^1.6",
+        "knplabs/knp-menu-bundle":"^2.2",
         "lexik/form-filter-bundle": "^5.0",
-        "craue/twigextensions-bundle": "^2.1",
-        "knplabs/knp-snappy-bundle": "^1.4",
+        "craue/twigextensions-bundle": "^2.4",
+        "knplabs/knp-snappy-bundle": "^1.6",
         "tomnomnom/etherpad-lite-client": "^0.1.2",
-        "zendframework/zend-soap": "^2.6",
-        "debril/rss-atom-bundle": "^3.0",
+        "zendframework/zend-soap": "^2.7",
+        "debril/rss-atom-bundle": "^4.2",
         "idci/color-scheme-bundle": "dev-master",
         "ci/restclientbundle": "^2.0",
-        "liip/imagine-bundle": "^1.6",
-        "nette/utils": "^2.4",
-        "ezyang/htmlpurifier": "^4.8",
-        "symfony/stopwatch": "^3.2",
-        "php-imap/php-imap": "^2.0",
-        "whiteoctober/breadcrumbs-bundle": "^1.2",
+        "liip/imagine-bundle": "^2.1",
+        "nette/utils": "^3.0",
+        "ezyang/htmlpurifier": "^4.1",
+        "symfony/stopwatch": "^3.4",
+        "php-imap/php-imap": "^3.0",
+        "whiteoctober/breadcrumbs-bundle": "^1.4",
         "friendsofsymfony/elastica-bundle": "^5.0.3",
-        "eluceo/ical": "^0.14.0",
-        "sabre/vobject": "^4.1",
+        "eluceo/ical": "^0.15",
         "sabre/dav": "^3.2.2",
         "mediawiki/oauthclient": "^0.1.2",
         "nategood/httpful": "^0.2.20",
-        "egulias/email-validator": "^2.1",
-        "friendsofsymfony/ckeditor-bundle": "^1.1"
+        "friendsofsymfony/ckeditor-bundle": "^1.2"
     },
     "require-dev": {
-        "sensio/generator-bundle": "^3.0",
-        "symfony/phpunit-bridge": "^3.0",
-        "doctrine/doctrine-fixtures-bundle": "^2.3",
-        "codeception/codeception": "^2.4"
+        "sensio/generator-bundle": "^3.1",
+        "symfony/phpunit-bridge": "^4.2",
+        "doctrine/doctrine-fixtures-bundle": "^3.1",
+        "codeception/codeception": "^2.5"
     },
     "scripts": {
         "symfony-scripts": [
